+++ Package.swift
@@ -1,0 +1,8 @@
+import PackageDescription
+
+/**
+    EasterDate package (Swift Package Manager)
+*/
+let package = Package(
+    name: "EasterDate"
+)