+++ Tests/LinuxMain.swift
@@ -1,0 +1,9 @@
+import XCTest
+@testable import EasterDateTestSuite
+
+/**
+    Test harness for Linux builds
+*/
+XCTMain([
+	 testCase(EasterDateTests.allTests),
+])