+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # EasterDate Change Log
 
+## 0.6.0 (2016-05-29)
+
+##### Enhancements
+
+* Add support for Swift Package Manager.
+* Add Linux target.
+
+##### Bug Fixes
+
+* None.
+
 ## 0.5.0 (2016-05-29)
 
 ##### Enhancements